requirements.txt
Django==3.1.7
pytz==2021.3
sqlparse==0.4.1
psycopg2>=2.8
redis>=3.5
celery>=5.0
asgiref==3.3.4
pycbrf==1.0.0
google-api-python-client==2.51.0
django_celery_beat==2.3.0
pytelegrambotapi==4.6.1